Update README.md
The "Simple Spelling Checker with Tkinter" project is a basic yet functional spelling checker implemented in Python. The project utilizes the NLTK (Natural Language Toolkit) library, a popular tool for natural language processing tasks. The primary objective of the project is to provide a graphical user interface (GUI) application that allows users to input text and receive real-time feedback on the correctness of the words used.